Added temporary changes to background, still working through this
