--- src/GameOver.css
@@ -1,24 +1,0 @@
-.game-over{
-    text-align: center;
-    vertical-align: center;
-    width: 500px;
-    padding-bottom: 30px;
-    margin: 0 auto;
-}
-.checkmate{
-    color: black;
-    font-size: 80px;
-    margin-bottom: 30px;
-    padding: 20px;
-    background-color:white;
-    border: 5px black solid;
-}
-button {
-    background-color:white;
-    color: black;
-    border: none;
-    padding: 10px 20px;
-    margin: 5px;
-    font-size: 16px;
-    cursor: pointer;
-}
